fix: index qualified type identity for castxml/clang public-contract closure
Codex review found the exact-qualified-reference fix (previous commit)
only actually worked for the DWARF backend convention, where
RecordType/EnumType.name itself already IS the qualified spelling.
_index_surface_types only ever keyed record_by_name/enum_by_name by
.name and its bare '::' tail -- never by .qualified_name -- so under the
castxml/clang convention (bare .name + separate .qualified_name), a
queued spelling equal to the qualified form never resolved anything,
leaving exact_type_identities empty regardless of how exact the
reference actually was. A second, related gap: _walk_type_closure only
ever added the bare .name to public_types, never the qualified spelling,
so even with indexing fixed, a qualified candidate would still never
appear in the public_types intersection _confirmed_type_matches checks.

Fixed both: _index_surface_types now also indexes by .qualified_name
(guarded against double-counting when qualified_name coincides with an
already-registered key, e.g. DWARF's unset qualified_name or an
unnamespaced castxml/clang record), and _walk_type_closure now
unconditionally adds each visited record/enum's own qualified spelling
to public_types too, matching the existing anti-hiding 'mark every match
public' rule for the bare name -- ambiguity-safety is still enforced
separately via ambiguous_type_names/exact_type_identities, not by
omission from public_types.

Confirmed via dumper_castxml.py's own _type_name that a real castxml-
produced signature never actually spells a record reference qualified
(the record's own "name" XML attribute is always local/bare, with
namespace context resolved separately) -- so the fp-rate corpus's
ambiguous_namespaced_leaf case, and any other real castxml-signature-only
scenario, remains and should remain UNKNOWN_UNRESOLVED regardless of this
fix. This closes the indexing/representation gap for whichever producer
does supply a genuinely qualified reference under that convention (a
hand-built Change via the public evaluate_change_contract_relevance API,
or a different producer/backend).

Adds test_exact_qualified_reference_now_confirms_castxml_convention and
a double-counting regression guard. Moving the growing
TestPublicModeAmbiguousTypeRoot class into its own file
(test_contract_evaluation_ambiguous_type_root.py) since
test_contract_evaluation.py had grown past the AI-readiness 2000-line
hard cap.

Updates the earlier test_a_public_header_sibling_no_public_api_reaches_
is_not_confirmed's public_types assertion, which is now a legitimately
larger (but still correctly non-confirming) set given the unconditional
qualified-spelling addition.

Verified: full targeted suite (contract evaluation, surface, exports,
replay, evidence collection, shadow measurement, export surface -- 913
tests) passes; scripts/check_fp_rate.py stays 0 FP/0 FN;
scripts/check_tier_accuracy.py unaffected; mypy/ruff clean;
scripts/check_ai_readiness.py back to 0 errors.
