+++ changelog.d/20260809_233000_claude_public_contract_exact_qualified_reference.md
@@ -13,4 +13,9 @@ A new changelog fragment. See changelog.d/README.md for the workflow.
   qualified identities were reached via an exact, unambiguous spelling
   (`exact_type_identities`), so `ns1::Point` confirms `IN_CONTRACT` while a
   type reached *only* through a shared, ambiguous bare tail (no side ever
-  spells the qualifier) is still correctly left unresolved.
+  spells the qualifier) is still correctly left unresolved. The closure's
+  own type index (`_index_surface_types`) now also keys on `RecordType`/
+  `EnumType.qualified_name`, not just the bare `.name` and its trailing
+  `::` segment, so this fix applies under the bare-name-plus-separate-
+  qualified-name convention (castxml/clang) as well as DWARF's
+  qualifier-baked-into-`.name` convention.
